# [pt][quant] Optimized qadd_scalar